pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "bluesky-scripts"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
description = "Bluesky plans and scripts"
dependencies = [
    "bluesky",
    "ophyd",
    "ophyd_async @ git+https://github.com/bluesky/ophyd-async@1020-pmac-with-gaps",
    "pandas",
    "matplotlib",
    "databroker",
    "scipy",
    "dls-dodal",
    "typer",
    "ipython"
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "copier",
    "pipdeptree",
    "pyright",
    "pytest",
    "pytest-cov",
    "ruff",
    "tox-direct",
    "types-mock",
]

[project.scripts]
imaging-bluesky = "bluesky_scripts.__main__:main"

[project.urls]
GitHub = "https://github.com/LuisFSegalla/BlueSkyTestPlans"

[[project.authors]] # Further authors may be added by duplicating this section
email = "Luis.Segalla@diamond.ac.uk"
name = "Luis Segalla"
